[script]: Improved [neonkube-nuget-dev.ps1] so you can now interrupt it with CRTL-C without messing up your [*.csproj] files.  Just rerun the script to restore them.

+++ ToolBin/neonkube-nuget-dev.ps1
@@ -92,7 +92,10 @@ function SetVersion
     $orgVersion     = $match.Groups[1].Value
     $tmpProjectFile = $orgProjectFile.Replace("<Version>$orgVersion</Version>", "<Version>$version</Version>")
 
-    Copy-Item "$projectPath" "$projectPath.bak"
+    if (!(Test-Path "$projectPath.bak"))
+    {
+        Copy-Item "$projectPath" "$projectPath.bak"
+    }
     
     $tmpProjectFile | Out-File -FilePath "$projectPath" -Encoding utf8
 }
